DEVOIR_CORRIGÉ/musicapi/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
promise==2.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==3.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0